+++ app/src/main/AndroidManifest.xml
@@ -19,7 +19,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name="ayds.newyork.songinfo.moredetails.presentation.view.OtherInfoWindow"
+        <activity android:name="ayds.newyork.songinfo.moredetails.presentation.view.MoreDetailsView"
             android:theme="@style/Theme.AppCompat.NoActionBar"/>
     </application>
