+++ app/src/main/java/ayds/newyork/songinfo/moredetails/data/external/nytimes/artistinfo/NYTimesArtistInfoServiceImpl.kt
@@ -1,0 +1,4 @@
+package ayds.newyork.songinfo.moredetails.data.external.nytimes.artistinfo
+
+class NYTimesArtistInfoServiceImpl {
+}
