+++ app/src/main/java/ayds/newyork/songinfo/moredetails/domain/repository/ArtistRepository.kt
@@ -1,0 +1,4 @@
+package ayds.newyork.songinfo.moredetails.domain.repository
+
+interface ArtistRepository {
+}
