--- app/src/main/java/ayds/newyork/songinfo/moredetails/model/data/ArtistRepositoryImpl.kt
@@ -1,6 +1,0 @@
-package ayds.newyork.songinfo.moredetails.model.data
-
-import ayds.newyork.songinfo.moredetails.model.domain.repository.ArtistRepository
-
-class ArtistRepositoryImpl : ArtistRepository {
-}
