--- app/src/main/java/ayds/newyork/songinfo/moredetails/model/data/external/nytimes/NYTimesArtistInfoService.kt
@@ -1,4 +1,0 @@
-package ayds.newyork.songinfo.moredetails.model.data.external.nytimes
-
-interface NYTimesArtistInfoService {
-}
