--- app/src/main/java/ayds/newyork/songinfo/moredetails/model/data/external/nytimes/artistinfo/NYTimesArtistInfoServiceImpl.kt
@@ -1,4 +1,0 @@
-package ayds.newyork.songinfo.moredetails.model.data.external.nytimes.artistinfo
-
-class NYTimesArtistInfoServiceImpl {
-}
