--- app/src/main/java/ayds/newyork/songinfo/moredetails/model/data/external/nytimes/artistinfo/NYTimesToArtistInfoResolver.kt
@@ -1,2 +1,0 @@
-package ayds.newyork.songinfo.moredetails.model.data.external.nytimes.artistinfo
-
