--- app/src/main/java/ayds/newyork/songinfo/moredetails/model/domain/repository/ArtistRepository.kt
@@ -1,4 +1,0 @@
-package ayds.newyork.songinfo.moredetails.model.domain.repository
-
-interface ArtistRepository {
-}
